model-training/requirements.txt
# model-training/requirements-local.txt
# Core dependencies only
pandas==2.1.0
numpy==1.24.3
python-dotenv==1.0.0
pyyaml==6.0.1

# GCP
google-cloud-storage==2.10.0
google-cloud-bigquery==3.11.4
google-cloud-aiplatform==1.36.0

# Model training
duckdb==0.9.0
sqlglot==19.0.0
plotly==5.17.0
mlflow==2.8.0

# Utils
colorama==0.4.6